# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillBridge AI",
    "description": "AI-powered career guidance for India's workforce",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dashboard/summary": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Summary",
        "operationId": "get_summary_dashboard_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UserDashboardResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ngo/beneficiaries": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Ngo Beneficiaries",
        "operationId": "get_ngo_beneficiaries_ngo_beneficiaries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ngo/outcomes": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Ngo Outcomes",
        "operationId": "get_ngo_outcomes_ngo_outcomes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ngo/skill-gaps": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Ngo Skill Gaps",
        "operationId": "get_ngo_skill_gaps_ngo_skill_gaps_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/govt/analytics": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Govt Analytics",
        "operationId": "get_govt_analytics_govt_analytics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/govt/placements": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Govt Placements",
        "operationId": "get_govt_placements_govt_placements_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/govt/skill-gaps": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Govt Skill Gaps",
        "operationId": "get_govt_skill_gaps_govt_skill_gaps_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/talent/matches": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Talent Matches",
        "operationId": "get_talent_matches_talent_matches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/gap-analysis/report": {
      "get": {
        "tags": [
          "Gap Analysis"
        ],
        "summary": "Get Report",
        "description": "Returns cached gap analysis report.\nRecomputes automatically if stale or missing.",
        "operationId": "get_report_gap_analysis_report_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_GapAnalysisReportResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gap-analysis/run": {
      "post": {
        "tags": [
          "Gap Analysis"
        ],
        "summary": "Force Run",
        "description": "Forces a fresh recompute regardless of cache state.\nCalled when user clicks 'Re-run Analysis'.",
        "operationId": "force_run_gap_analysis_run_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_GapAnalysisReportResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gap-analysis/roadmap": {
      "get": {
        "tags": [
          "Gap Analysis"
        ],
        "summary": "Get Roadmap",
        "description": "Returns only the roadmap portion of the current report.",
        "operationId": "get_roadmap_gap_analysis_roadmap_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse"
      },
...
      "APIResponse_GapAnalysisReportResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GapAnalysisReportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[GapAnalysisReportResponse]"
      },
...
      "APIResponse_UserDashboardResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserDashboardResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[UserDashboardResponse]"
      },
...
      "DashboardLocationSchema": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          }
        },
        "type": "object",
        "title": "DashboardLocationSchema"
      },
      "DashboardUserSchema": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "user_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Type"
          },
          "preferred_lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Lang",
            "default": "en"
          }
        },
        "type": "object",
        "title": "DashboardUserSchema"
      },
...
      "GapAnalysisReportResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "target_role": {
            "type": "string",
            "title": "Target Role"
          },
          "overall_match_score": {
            "type": "number",
            "title": "Overall Match Score",
            "default": 0.0
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/SkillGapItem"
            },
            "type": "array",
            "title": "Gaps"
          },
          "roadmap": {
            "items": {
              "$ref": "#/components/schemas/RoadmapStep"
            },
            "type": "array",
            "title": "Roadmap"
          },
          "motivational_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivational Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "from_cache": {
            "type": "boolean",
            "title": "From Cache",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "target_role",
          "gaps",
          "roadmap",
          "created_at"
        ],
        "title": "GapAnalysisReportResponse",
        "description": "Full gap analysis report."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoadmapStep": {
        "properties": {
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "estimated_duration": {
            "type": "string",
            "title": "Estimated Duration"
          },
          "resources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Resources",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "order",
          "title",
          "description",
          "estimated_duration"
        ],
        "title": "RoadmapStep",
        "description": "Single step in a learning roadmap."
      },
...
      "SkillGapItem": {
        "properties": {
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "current_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "required_level": {
            "$ref": "#/components/schemas/SkillLevel"
          },
          "importance": {
            "type": "number",
            "title": "Importance",
            "default": 1.0
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "skill_name",
          "required_level",
          "description"
        ],
        "title": "SkillGapItem",
        "description": "Specific skill gap identified."
      },
...
      "SkillLevel": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced",
          "expert"
        ],
        "title": "SkillLevel"
      },
...
      "UserDashboardResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/DashboardUserSchema"
          },
          "profile_completion_pct": {
            "type": "integer",
            "title": "Profile Completion Pct",
            "default": 0
          },
          "onboarding_done": {
            "type": "boolean",
            "title": "Onboarding Done",
            "default": false
          },
          "quick_assessment_done": {
            "type": "boolean",
            "title": "Quick Assessment Done",
            "default": false
          },
          "gap_analysis_done": {
            "type": "boolean",
            "title": "Gap Analysis Done",
            "default": false
          },
          "gap_analysis_stale": {
            "type": "boolean",
            "title": "Gap Analysis Stale",
            "default": false
          },
          "last_assessment_at": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Last Assessment At"
          },
          "extracted_skills": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Extracted Skills",
            "default": []
          },
          "career_interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Career Interests",
            "default": []
          },
          "location": {
            "$ref": "#/components/schemas/DashboardLocationSchema"
          },
          "job_matches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Job Matches",
            "default": []
          },
          "recommended_courses": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recommended Courses",
            "default": []
          },
          "role_specific": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Specific"
          },
          "ai_highlight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Highlight"
          },
          "recent_activity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Activity",
            "default": []
          },
          "progress_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Progress Summary",
            "default": {
              "courses_completed": 0,
              "assessments_taken": 0,
              "skills_verified": 0
            }
          },
          "notifications_count": {
            "type": "integer",
            "title": "Notifications Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "user",
          "location"
        ],
        "title": "UserDashboardResponse",
        "description": "Dashboard summary for the user.",
        "example": {
          "ai_highlight": "Focus on Python to unlock 5 new roles.",
          "career_interests": [
            "Software Development"
          ],
          "extracted_skills": [],
          "gap_analysis_done": false,
          "gap_analysis_stale": false,
          "job_matches": [],
          "location": {
            "city": "Faizabad",
            "state": "uttar_pradesh"
          },
          "notifications_count": 0,
          "onboarding_done": true,
          "profile_completion_pct": 87,
          "progress_summary": {
            "assessments_taken": 1,
            "courses_completed": 2,
            "skills_verified": 3
          },
          "quick_assessment_done": false,
          "user": {
            "name": "Mohan Blue",
            "preferred_lang": "en",
            "user_type": "individual_youth"
          }
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}